+++ Changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- [Object object] error message when device enumeration failed.
+### Added
+- Function `logError` to ease logging error.
+
 ## 5.8.0 - 2021-11-08
 ### Added
 - Chevron to `DeviceSelector`